requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
constantly==15.1.0
cryptography==38.0.4
cycler==0.11.0
daphne==4.0.0
decorator==5.1.1
Django==3.2.16
firebase-admin==6.0.1
fonttools==4.38.0
gcloud==0.17.0
googleapis-common-protos==1.57.0
greenlet==2.0.1
httplib2==0.21.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.1.0
incremental==22.10.0
jws==0.1.3
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.5.3
msgpack==1.0.4
numpy==1.21.6
oauth2client==3.0.0
obspy==1.3.1
packaging==22.0
Pillow==9.3.0
protobuf==4.21.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.4.3
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2022.6
redis==4.4.0
requests-toolbelt==0.7.0
rsa==4.9
scipy==1.7.3
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
zipp==3.11.0
zope.interface==5.5.2
